+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Hide condition fields in replication settings when `each_n` and `each_s` are null or undefined, [PR-106](https://github.com/reductstore/web-console/pull/106)
 - Use streams for downloading large records, [PR-113](https://github.com/reductstore/web-console/pull/113)
+- Handle .mcap files with correct content type on upload and download, [PR-114](https://github.com/reductstore/web-console/pull/114)
 
 ### Fixed